Create Twist & SubGroup
Here is where we create the Twist Attack on other curves. What it does is to check whether the given public key X falls on the given twisted curve. If you receive errors, check the curve errors and change or remove it. Because not all public key falls on the same twisted curve, so you might need to increase or decrease the values